feat: 예약 정보가 있는 경우 테마 수정 / 삭제 불가능 기능 추가

+++ src/main/java/web/theme/service/ThemeService.java
@@ -2,8 +2,11 @@ package web.theme.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import web.exception.ErrorCode;
+import web.reservation.repository.ReservationRepository;
 import web.theme.dto.ThemeRequestDto;
 import web.theme.dto.ThemeResponseDto;
+import web.theme.exception.ThemeException;
 import web.theme.repository.ThemeRepository;
 
 import java.util.List;
@@ -14,6 +17,7 @@ import java.util.stream.Collectors;
 public class ThemeService {
 
     private final ThemeRepository themeRepository;
+    private final ReservationRepository reservationRepository;
 
     public long save(ThemeRequestDto themeRequestDto) {
         return themeRepository.save(themeRequestDto.toEntity());
@@ -27,6 +31,14 @@ public class ThemeService {
     }
 
     public void deleteById(long themeId) {
+        if (isReservationExist(themeId)) {
+            throw new ThemeException(ErrorCode.RESERVATION_EXIST);
+        }
         themeRepository.delete(themeId);
+    }
+
+    private boolean isReservationExist(long themeId) {
+        return !reservationRepository.findAllByThemeId(themeId)
+                .isEmpty();
     }
 }
